Updated local-work-size and shared-memory tests to only run on CUDA backend
